+++ src/i18n/zh-tw.ts
@@ -1,19 +1,19 @@
 import { LocalProperty } from "./types";
 
 export const ZHtw: LocalProperty = {
-  "Tab_Title": "自定義圖標",
-  "Tab_SidebarIcons": "側邊工作區圖標",
-  "SidebarIcons_FileName": "輸入檔案名稱",
-  "Tab_FolderIcons": "資料夾圖標",
-  "FolderIcons_FileName": "輸入資料夾路徑",
-  "Tab_About": "關於",
-  "IconID": "{num}",
-  "IconImg": "輸入圖標路徑",
-  "Type_Custom": "自定義",
-  "Type_Lucide": "lucide",
-  "RemoveIconButton": "移除",
-  "AddIconButton": "新增圖標",
-  "About": 
+  Tab_Title: "自定義圖標",
+  Tab_SidebarIcons: "側邊工作區圖標",
+  SidebarIcons_FileName: "輸入檔案名稱",
+  Tab_FolderIcons: "資料夾圖標",
+  FolderIcons_FileName: "輸入資料夾路徑",
+  Tab_About: "關於",
+  IconID: "{num}",
+  IconImg: "輸入圖標路徑",
+  Type_Custom: "自定義",
+  Type_Lucide: "lucide",
+  RemoveIconButton: "移除",
+  AddIconButton: "新增圖標",
+  About: 
   `
   <h5>簡要說明</h5>
   <ul>