+++ src/i18n/zh.ts
@@ -17,8 +17,8 @@ export const ZH: LocalProperty = {
   `
   <h5>简要说明</h5>
   <ul>
-    <li>自定义图标支持: 在线URL 或 本地相对(绝对)文件路径 或 base64编码 或 svg标签。 详情查看<a herf="https://github.com/RavenHogWarts/obsidian-custom-sidebar-icons/blob/master/README_ZH.md" target="_blank">README</a>。</li>
-    <li>lucide源图标支持直接使用lucide图标名称。 详情查看<a herf="https://lucide.dev/icons/" target="_blank">lucide</a>。</li>
+    <li>自定义图标支持: 在线URL 或 本地相对(绝对)文件路径 或 base64编码 或 svg标签。 详情查看<a href="https://github.com/RavenHogWarts/obsidian-custom-sidebar-icons/blob/master/README_ZH.md" target="_blank">README</a>。</li>
+    <li>lucide源图标支持直接使用lucide图标名称。 详情查看<a href="https://lucide.dev/icons/" target="_blank">lucide</a>。</li>
   </ul>
   `
 };